+++ Examples/Swift/DeskObserver/DeskObserver/ViewController.swift
@@ -50,61 +50,55 @@ class ViewController: UIViewController {
         
         // There's just a one Proximity Observer object used to configure reactions to proximity changes.
         // The configuration of a zone consists of defining a desired approximate range (0.5m or 1.5m in this example), 
-        // attachment key, value pair ("desk" : "blueberry", "desk" : "mint" or "venue" : "office" in this example)
+        // a tag ("blueberry", "mint" and "office" in this example)
         // and a handler block called when a zone "enter" / "exit" / "change of beacons in range" happens.
         //
-        // For more info on attachment-based beacon identification, check out the README in this repository
+        // For more info on tag-based beacon identification, check out the README in this repository
         // ( https://github.com/Estimote/iOS-Proximity-SDK/blob/master/README.md ).
         
         // TODO: The demo assumes proper settings configuration in Estimote Cloud. Make sure that:
         // - you have at least 2 beacons assigned to your Estimote Cloud account
-        // - one of the beacons has an attachment `{"desk":"blueberry","venue":"office"}`
-        // - another beacon has an attachment `{"desk":"mint","venue":"office"}`
+        // - one of the beacons has tags "blueberry" and "office"
+        // - another beacon has tags "mint" and "office
         
         self.proximityObserver = EPXProximityObserver(credentials: credentials, errorBlock: { error in
             print("Ooops! \(error)")
         })
         
-        let blueberryZone = EPXProximityZone(range: EPXProximityRange.custom(desiredMeanTriggerDistance: 0.5)!,
-                                             attachmentKey: "desk",
-                                             attachmentValue: "blueberry")
-        blueberryZone.onEnterAction = { attachment in
+        let blueberryZone = EPXProximityZone(range: EPXProximityRange.near,
+                                             tag: "test")
+        blueberryZone.onEnterAction = { zoneContext in
             print("Enter blueberry (close range)")
             self.blueberryDeskLabel.backgroundColor = self.blueberryColor
             self.blueberryDeskLabel.textColor = UIColor.white
         }
-        blueberryZone.onExitAction = { attachment in
+        blueberryZone.onExitAction = { zoneContext in
             print("Exit blueberry (close range)")
             self.blueberryDeskLabel.backgroundColor = UIColor.white
             self.blueberryDeskLabel.textColor = self.blueberryColor
         }
         
         let mintZone = EPXProximityZone(range: EPXProximityRange.custom(desiredMeanTriggerDistance: 0.5)!,
-                                        attachmentKey: "desk",
-                                        attachmentValue: "mint")
-        mintZone.onEnterAction = { attachment in
+                                        tag: "mint")
+        mintZone.onEnterAction = { zoneContext in
             print("Enter mint (close range)")
             self.mintDeskLabel.backgroundColor = self.mintColor
             self.mintDeskLabel.textColor = UIColor.white
         }
-        mintZone.onExitAction = { attachment in
+        mintZone.onExitAction = { zoneContext in
             print("Exit mint (close range)")
             self.mintDeskLabel.backgroundColor = UIColor.white
             self.mintDeskLabel.textColor = self.mintColor
         }
         
         let closeVenueZone = EPXProximityZone(range: EPXProximityRange.custom(desiredMeanTriggerDistance: 0.5)!,
-                                              attachmentKey: "venue",
-                                              attachmentValue: "office")
-        closeVenueZone.onChangeAction = { attachmentsInside in
-            print("Currently, there are \(attachmentsInside.count) attachments in close range:")
-            print("\(attachmentsInside.map({ $0.payload.description }).joined(separator: "\n"))")
-            print("")
+                                              tag: "office")
+        closeVenueZone.onChangeAction = { zoneContextsInside in
+            print("Currently, there are \(zoneContextsInside.count) attachments in close range:")
         }
         
         let midVenueZone = EPXProximityZone(range: EPXProximityRange.custom(desiredMeanTriggerDistance: 1.5)!,
-                                            attachmentKey: "venue",
-                                            attachmentValue: "office")
+                                            tag: "office")
         midVenueZone.onEnterAction = { attachment in
             print("Enter venue (mid range)")
             self.venueLabel.backgroundColor = self.venueColor
@@ -115,10 +109,8 @@ class ViewController: UIViewController {
             self.venueLabel.backgroundColor = UIColor.white
             self.venueLabel.textColor = self.venueColor
         }
-        midVenueZone.onChangeAction = { attachmentsInside in
-            print("Currently, there are \(attachmentsInside.count) attachments in mid range:")
-            print("\(attachmentsInside.map({ $0.payload.description }).joined(separator: "\n"))")
-            print("")
+        midVenueZone.onChangeAction = { zoneContextsInside in
+            print("Currently, there are \(zoneContextsInside.count) attachments in mid range: \(zoneContextsInside.map({($0 as! ProximityZoneContext).deviceIdentifier}))")
         }
         
         self.proximityObserver.startObserving([mintZone, blueberryZone, closeVenueZone, midVenueZone])